pyproject.toml
[project]
name = "telegram-client-copier"
version = "0.1.0"
description = "Multi-tenant Telegram channel copier with admin controls."
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "fastapi>=0.110.0",
  "motor>=3.4.0",
  "bcrypt>=4.0.0",
  "email-validator>=2.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "telethon>=1.34.0",
  "typer>=0.12.0",
  "uvicorn>=0.29.0",
]

[project.scripts]
tg-copier = "app.main:run"

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
asyncio_mode = "auto"

